videos and posts avtorize

+++ controllers/post/removePost.js
@@ -1,8 +1,9 @@
 const {Post} = require('../../models/post')
 
 const removePost = async (req, res) => {
+   const { _id } = req.user;
    const {postId} = req.params; 
-   const removePost = await Post.findByIdAndRemove(postId);
+   const removePost = await Post.findByIdAndRemove({owner: _id, _id: postId});
    if(removePost) {
    res.status(200).json({ message: 'contact deleted' });   
    } else {
